requirements.txt
fastapi>=0.135,<1.0
uvicorn>=0.41,<1.0
openai>=2.28,<3.0
psycopg2-binary>=2.9,<3.0
sqlmodel>=0.0.37,<0.1
SQLAlchemy>=2.0,<3.0
pydantic>=2.12,<3.0
passlib>=1.7,<2.0
argon2-cffi>=25,<26
PyJWT>=2.12,<3.0
